refs #8 色々リファクタリング
