--- react_app/src/components/App.tsx
@@ -1,7 +1,0 @@
-import * as React from "react";
-
-export class App extends React.Component {
-  render() {
-    return <h1>Hello React!</h1>;
-  }
-}
