+++ react_app/src/components/TopPage/TopPage.tsx
@@ -1,0 +1,8 @@
+import * as React from 'react';
+
+/**
+ * トップページ
+ */
+export const TopPage = () => {
+  return <h1>Hello React!</h1>;
+}
