--- react_app/src/components/WithoutAppBarPage.tsx
@@ -1,7 +1,0 @@
-import * as React from "react";
-
-export class WithoutAppBarPage extends React.Component {
-  render() {
-    return <h1>このページはapp Barに含まれません</h1>;
-  }
-}
