+++ react_app/src/components/others/NotFound.tsx
@@ -1,0 +1,8 @@
+import * as React from 'react';
+
+/**
+ * ルーティング外のアクセスが来たときに表示させるNotFoundページ
+ */
+export const NotFound = () => {
+  return <h1>ページが見つかりません</h1>
+}
