+++ react_app/src/index.tsx
@@ -5,14 +5,14 @@ import * as ReactDOM from 'react-dom';
 // apollo
 import { ApolloProvider } from 'react-apollo';
 import { client } from './apolloClient'
-import { Routes } from './routes'
+import { Base } from './components/Base'
 
 // favicon 設定はindex.htmlで行なっている
 import './favicon.ico';
 
 ReactDOM.render(
   <ApolloProvider client={client}>
-    <Routes />
+    <Base />
   </ApolloProvider>,
   document.getElementById('app')
 );
